Further improvements to image seeding

+++ database/seeds/ImageSeeder.php
@@ -17,20 +17,25 @@ class ImageSeeder extends Seeder
         $avatar = new Avatar();
         $users = \App\User::all();
         foreach ($users as $user) {
-            $avatar->create($user->name)->save(public_path('images/avatars/' . $user->id . '.png'), $quality = 90);
+            $avatar->create($user->name)->save(storage_path('app/public/avatars/user-' . $user->id . '.png'), $quality = 90);
         }
         $numberOfProductImages = 100;
         $this->fetchAllProductImages($numberOfProductImages);
     }
 
     protected function fetchProductImage($id) {
-        $url = "https://loremflickr.com/640/480/vegetables";
-        $info = pathinfo($url);
-        $contents = file_get_contents($url);
-        $file = '/tmp/' . $info['basename'];
-        file_put_contents($file, $contents);
-        $uploaded_file = new UploadedFile($file, $info['basename']);
-        Storage::putFileAs('public/images/products/', $uploaded_file, $id.'.jpg');
+        $path = 'public/products';
+        $filename = 'product-'.$id.'.jpg';
+
+        if (! Storage::disk('local')->exists($path.'/'.$filename)) {
+            $url = "https://loremflickr.com/640/480/vegetables";
+            $info = pathinfo($url);
+            $contents = file_get_contents($url);
+            $file = '/tmp/' . $info['basename'];
+            file_put_contents($file, $contents);
+            $uploaded_file = new UploadedFile($file, $info['basename']);
+            Storage::putFileAs($path, $uploaded_file, $filename);
+        }
     }
 
     protected function fetchAllProductImages($count) {
